Centralize test utilities and shared components
This commit centralizes the test utilities and shared components to improve the test suite's organization and maintainability.

- A `tests/shared` directory was created to store reusable test utilities and components.
- `tests/fixtures/test_helpers.py` was moved to `tests/shared/helpers.py`.
- A `tests/shared/mocks.py` file was created, and all the mock classes from `tests/integration/conftest.py` were moved to it.
- All imports were updated to reflect the new file structure.
