+++ tests/integration/agents/test_agent_lifecycle.py
@@ -22,7 +22,7 @@ from agents.software_engineer.workflows.sequential_workflows import (
 )
 
 # Test utilities
-from tests.fixtures.test_helpers import (
+from tests.shared.helpers import (
     create_mock_llm_client,
     create_mock_session_state,
     create_test_workspace,
