+++ tests/integration/agents/test_single_agent_patterns.py
@@ -15,7 +15,7 @@ from agents.devops.devops_agent import MyDevopsAgent
 from agents.software_engineer.agent import root_agent as software_engineer_agent
 
 # Test utilities
-from tests.fixtures.test_helpers import (
+from tests.shared.helpers import (
     create_mock_llm_client,
     create_mock_session_state,
     create_test_workspace,
