+++ tests/integration/tools_system/test_system_instruction_compliance.py
@@ -19,7 +19,7 @@ from agents.software_engineer.sub_agents.documentation.agent import documentatio
 from agents.software_engineer.sub_agents.testing.agent import testing_agent
 
 # Test utilities
-from tests.fixtures.test_helpers import (
+from tests.shared.helpers import (
     create_mock_llm_client,
     create_mock_session_state,
     create_test_workspace,
